requirements.txt
flask==3.0.2
python-dotenv==1.0.1
google-generativeai==0.3.2
requests==2.31.0
python-jose==3.3.0
flask-cors==4.0.0 